gh #190 Updated L3 test folder structure
