+++ host/tests/L3_TestCases/dsVideoPort/dsVideoPort_L3_testSetup.yml
@@ -1,0 +1,37 @@
+#** *****************************************************************************
+# *
+# * If not stated otherwise in this file or this component's LICENSE file the
+# * following copyright and licenses apply:
+# *
+# * Copyright 2024 RDK Management
+# *
+# * Licensed under the Apache License, Version 2.0 (the "License");
+# * you may not use this file except in compliance with the License.
+# * You may obtain a copy of the License at
+# *
+# *
+# http://www.apache.org/licenses/LICENSE-2.0
+# *
+# * Unless required by applicable law or agreed to in writing, software
+# * distributed under the License is distributed on an "AS IS" BASIS,
+# * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# * See the License for the specific language governing permissions and
+# * limitations under the License.
+# *
+#* ******************************************************************************
+
+# this file is generated by the testing suites, when a new test is modified, this configuration file will be filled out and updated.
+#location: https://github.com/rdkcentral/rdk-halif-test-device_settings/blob/develop/host/tests/L3_TestCases/dsVideoPort
+
+dsVideoPort:  # Prefix must always exist
+  description: "dsVideoPort Device Settings test setup"
+  assets:
+    device:
+      test1_Verify_Video_Content_Formats_Callbacks:
+        artifacts: #list of artifacts to be copied to device
+        execute: #list of commands to be executed on device
+        streams: #list of streams to be copied to device
+      test2_Check_DisplayConnected:
+        artifacts:
+        execute:
+        streams:
