+++ host/tests/L3_TestCases/dsVideoPort/dsVideoPort_test2_EnableDisableAndVerifyPortStatus.py
@@ -26,7 +26,7 @@ import sys
 
 # Get directory path and append to system path
 dir_path = os.path.dirname(os.path.realpath(__file__))
-sys.path.append(dir_path+"/../")
+sys.path.append(dir_path+"/../../")
 
 # Import required classes from modules
 from dsClasses.dsVideoPort import dsVideoPortClass
