+++ host/tests/L3_TestCases/dsVideoPort/dsVideoPort_test9_VerifyBackgroundColor.py
@@ -23,10 +23,11 @@
 
 import os
 import sys
+from enum import Enum, auto
 
 # Append the parent directory to system path for module imports
 dir_path = os.path.dirname(os.path.realpath(__file__))
-sys.path.append(dir_path+"/../")
+sys.path.append(dir_path+"/../../")
 
 # Import required classes from modules
 from dsClasses.dsVideoPort import dsVideoPortClass
@@ -35,19 +36,25 @@ from raft.framework.plugins.ut_raft.configRead import ConfigRead
 from raft.framework.plugins.ut_raft.utPlayer import utPlayer
 from raft.framework.plugins.ut_raft.utUserResponse import utUserResponse
 
-class dsVideoPort_test5_VerifyHDCP_Callback(utHelperClass):
+class dsVideoBackgroundColor(Enum):
+    dsVIDEO_BGCOLOR_BLUE = 0   # Background color BLUE
+    dsVIDEO_BGCOLOR_BLACK = 1  # Background color BLACK
+    dsVIDEO_BGCOLOR_NONE = 2   # Background color NONE
+    dsVIDEO_BGCOLOR_MAX = 3    # Out of range
+
+class dsVideoPort_test9_VerifyBackgroundColor(utHelperClass):
     """
-    Test class for verifying HDCP callback during HDMI plug/unplug.
+    Test class for verifying the background color of video ports.
     """
 
-    testName  = "test5_VerifyHDCP_Callback"
+    testName  = "test9_VerifyBackgroundColor"
     testSetupPath = dir_path + "/dsVideoPort_L3_testSetup.yml"
     moduleName = "dsVideoPort"
     rackDevice = "dut"
 
     def __init__(self):
         """
-        Initializes the test5_VerifyHDCP_Callback test with necessary setup.
+        Initializes the test9_VerifyBackgroundColor test .
 
         Args:
             None.
@@ -57,18 +64,10 @@ class dsVideoPort_test5_VerifyHDCP_Callback(utHelperClass):
         # Test Setup configuration file
         self.testSetup = ConfigRead(self.testSetupPath, self.moduleName)
 
-        # Open Session for player
-        #self.player_session = self.dut.getConsoleSession("ssh_player")
-
         # Open Session for hal test
         self.hal_session = self.dut.getConsoleSession("ssh_hal_test")
 
-        #player = self.cpe.get("test").get("player")
-
-        # Create player Class
-        #self.testPlayer = utPlayer(self.player_session, player)
-
-         # Create user response Class
+        # Create user response Class
         self.testUserResponse = utUserResponse()
 
         # Get path to device profile file
@@ -88,12 +87,12 @@ class dsVideoPort_test5_VerifyHDCP_Callback(utHelperClass):
         self.deviceDownloadPath = self.cpe.get("target_directory")
 
         #download test artifacts to device
-        url = self.testSetup.assets.device.test5_VerifyHDCP_Callback.artifacts
+        url = self.testSetup.assets.device.test9_VerifyBackgroundColor.artifacts
         if url is not None:
             self.downloadToDevice(url, self.deviceDownloadPath, self.rackDevice)
 
         #download test streams to device
-        url = self.testSetup.assets.device.test5_VerifyHDCP_Callback.streams
+        url = self.testSetup.assets.device.test9_VerifyBackgroundColor.streams
         if url is not None:
             self.downloadToDevice(url, self.deviceDownloadPath, self.rackDevice)
             for streampath in url:
@@ -117,15 +116,15 @@ class dsVideoPort_test5_VerifyHDCP_Callback(utHelperClass):
         """
 
         #Run test specific commands
-        cmds = self.testSetup.assets.device.test5_VerifyHDCP_Callback.execute
+        cmds = self.testSetup.assets.device.test9_VerifyBackgroundColor.execute
         if cmds is not None:
             for cmd in cmds:
                 self.writeCommands(cmd)
 
     #TODO: Current version supports only manual verification.
-    def testUnplugHDMI(self, unplug:True, manual=False):
+    def testVerifyBackGroundColor(self, manual=False,backgroundColor:str=0):
         """
-        Verifies whether the Video&audio displayed or not.
+        Verifies the HDCP Version .
 
         Args:
             manual (bool, optional): Manual verification (True: manual, False: other verification methods).
@@ -135,87 +134,60 @@ class dsVideoPort_test5_VerifyHDCP_Callback(utHelperClass):
             bool
         """
         if manual == True:
-            if unplug == True:
-                return self.testUserResponse.getUserYN("UnPlug the HDMI Cable? (Y/N):")
-            else :
-                return self.testUserResponse.getUserYN("Plug the HDMI Cable? (Y/N):")
+            return self.testUserResponse.getUserYN(f'is {backgroundColor} displayed on Analyzer (Y/N): ')
         else :
             #TODO: Add automation verification methods
             return False
 
-    def find_HDCPStatus(self, input_str: str, status: str) -> bool:
-        """
-        Finds HDCP status in a given input string.
-
-        Args:
-            input_str (str): The input log containing HDCP status.
-            status (str): The HDCP status string to look for.
-
-        Returns:
-            bool: True if the status is found, False otherwise.
-        """
-        if status in input_str:
-            return True
-        return False
-
-    def testEnablePortAndVerifyHDCP(self, port, index):
-        """
-        Enables a video port, performs HDMI plug/unplug actions, and verifies HDCP callbacks.
-
-        Args:
-            port (str): The video port name.
-            index (int): The port index.
-
-        Returns:
-            bool: Result of the HDCP verification.
-        """
-        self.log.stepStart(f'Enable {port} Port')
-        self.testdsVideoPort.enablePort(port, index)
-
-        # Enable HDCP for the port
-        self.testdsVideoPort.enable_HDCP(port, index)
-
-        # Verify unplug HDMI event
-        result = self.testUnplugHDMI(unplug=True, manual=True)
-        result = self.testdsVideoPort.read_Callbacks("HDCP Status Callback dsHdcpStatus_t:")
-        if self.find_HDCPStatus(result, "dsHDCP_STATUS_UNPOWERED"):
-            self.log.stepResult(True, "HDMI Unplug Callback found")
-        else:
-            self.log.stepResult(False, "HDMI Unplug Callback Not found")
-
-        # Verify plug HDMI event
-        result = self.testUnplugHDMI(unplug=False, manual=True)
-        result = self.testdsVideoPort.read_Callbacks("HDCP Status Callback dsHdcpStatus_t:")
-        if self.find_HDCPStatus(result, "dsHDCP_STATUS_AUTHENTICATED"):
-            self.log.stepResult(True, "HDMI Plug Callback found")
-        else:
-            self.log.stepResult(False, "HDMI Plug Callback Not found")
-
-        return result
-
     def testFunction(self):
-        """
-        Main test function to verify HDCP status callbacks during HDMI plug/unplug events.
+        """This function will test the Video Ports by enabling and disabling the ports
 
         Returns:
-            bool: Final result of the test.
+            bool
         """
+
+        # Download the assets listed in test setup configuration file
         self.testDownloadAssets()
+
+        # Run Prerequisites listed in the test setup configuration file
         self.testRunPrerequisites()
 
-        # Initialize the dsVideoPort class
+        # Create the dsVideoPort class
         self.testdsVideoPort = dsVideoPortClass(self.deviceProfile, self.hal_session)
+
+        self.log.testStart("test9_VerifyBackgroundColor", '1')
+
+        # Initialize the dsVideoPort module
         self.testdsVideoPort.initialise()
 
-        # Loop through all supported video ports and verify HDCP callbacks
-        for port, index in self.testdsVideoPort.getSupportedPorts():
-            result = self.testEnablePortAndVerifyHDCP(port, index)
+        # Loop through the supported Video ports
+        for port,index in self.testdsVideoPort.getSupportedPorts():
+            self.log.stepStart(f'Enable {port} Port')
+            self.log.step(f'Enable {port} Port')
 
+            # Enable the Video port
+            self.testdsVideoPort.enablePort(port, index)
+
+            # Enable the HDCP only for source devices
+            if self.testdsVideoPort.getDeviceType():
+                self.testdsVideoPort.enable_HDCP(port, index)
+
+            for backgroundColor in list(dsVideoBackgroundColor):
+                if backgroundColor != dsVideoBackgroundColor.dsVIDEO_BGCOLOR_MAX:
+                    self.testdsVideoPort.select_BackgroundColor(port, index, dsVideoBackgroundColor(backgroundColor).name)
+                    result = self.testVerifyBackGroundColor(True,dsVideoBackgroundColor(backgroundColor).name)
+
+            self.log.stepResult(result, "All parameters verified using HDMI Analyzer")
+
+
+        # Clean the assets downloaded to the device
         self.testCleanAssets()
+
+        # Delete the dsVideoPort class
         del self.testdsVideoPort
 
         return result
 
 if __name__ == '__main__':
-    test = dsVideoPort_test5_VerifyHDCP_Callback()
+    test = dsVideoPort_test9_VerifyBackgroundColor()
     test.run(False)
