+++ host/tests/dsClasses/dsVideoPort_test_suite.yml
@@ -1,3 +1,25 @@
+#** *****************************************************************************
+# *
+# * If not stated otherwise in this file or this component's LICENSE file the
+# * following copyright and licenses apply:
+# *
+# * Copyright 2024 RDK Management
+# *
+# * Licensed under the Apache License, Version 2.0 (the "License");
+# * you may not use this file except in compliance with the License.
+# * You may obtain a copy of the License at
+# *
+# *
+# http://www.apache.org/licenses/LICENSE-2.0
+# *
+# * Unless required by applicable law or agreed to in writing, software
+# * distributed under the License is distributed on an "AS IS" BASIS,
+# * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# * See the License for the specific language governing permissions and
+# * limitations under the License.
+# *
+#* ******************************************************************************
+
 # this file is generated by the testing suites, when a new test is modified, this configuration file will be filled out and updated.
 
 dsVideoPort:  # Prefix must always exist
@@ -9,9 +31,102 @@ dsVideoPort:  # Prefix must always exist
             0:
                 name: "L1 dsVideoPort - Sink"
                 tests:
+                    - "dsVideoPortInit_L1_positive"
+                    - "dsVideoPortInit_L1_negative"
+                    - "dsVideoPortTerm_L1_positive"
+                    - "dsVideoPortTerm_L1_negative"
+                    - "dsGetVideoPort_L1_positive"
+                    - "dsGetVideoPort_L1_negative"
+                    - "dsIsVideoPortEnabled_L1_positive"
+                    - "dsIsVideoPortEnabled_L1_negative"
+                    - "dsIsDisplayConnected_L1_positive"
+                    - "dsIsDisplayConnected_L1_negative"
+                    - "dsIsDisplaySurround_L1_positive"
+                    - "dsIsDisplaySurround_L1_negative"
+                    - "dsGetSurroundMode_L1_positive"
+                    - "dsGetSurroundMode_L1_negative"
+                    - "dsVideoFormatUpdateRegisterCB_L1_positive"
+                    - "dsVideoFormatUpdateRegisterCB_L1_negative"
+                    - "dsIsVideoPortActive_L1_positive"
+                    - "dsIsVideoPortActive_L1_negative"
+                    - "dsEnableHDCP_L1_positive"
+                    - "dsEnableHDCP_L1_negative"
+                    - "dsIsHDCPEnabled_L1_positive"
+                    - "dsIsHDCPEnabled_L1_negative"
+                    - "dsEnableVideoPort_L1_positive"
+                    - "dsEnableVideoPort_L1_negative"
+                    - "dsSetResolution_L1_positive"
+                    - "dsSetResolution_L1_negative"
+                    - "dsGetResolution_L1_positive"
+                    - "dsGetResolution_L1_negative"
+                    - "dsRegisterHdcpStatusCallback_L1_positive"
+                    - "dsRegisterHdcpStatusCallback_L1_negative"
+                    - "dsGetHDCPStatus_L1_positive"
+                    - "dsGetHDCPStatus_L1_negative"
+                    - "dsGetHDCPProtocol_L1_positive"
+                    - "dsGetHDCPProtocol_L1_negative"
+                    - "dsGetHDCPReceiverProtocol_L1_positive"
+                    - "dsGetHDCPReceiverProtocol_L1_negative"
+                    - "dsGetHDCPCurrentProtocol_L1_positive"
+                    - "dsGetHDCPCurrentProtocol_L1_negative"
+                    - "dsGetTVHDRCapabilities_L1_positive"
+                    - "dsGetTVHDRCapabilities_L1_negative"
+                    - "dsSupportedTvResolutions_L1_positive"
+                    - "dsSupportedTvResolutions_L1_negative"
+                    - "dsSetForceDisable4KSupport_L1_positive"
+                    - "dsSetForceDisable4KSupport_L1_negative"
+                    - "dsGetForceDisable4KSupport_L1_positive"
+                    - "dsGetForceDisable4KSupport_L1_negative"
+                    - "dsGetVideoEOTF_L1_positive"
+                    - "dsGetVideoEOTF_L1_negative"
+                    - "dsGetMatrixCoefficients_L1_positive"
+                    - "dsGetMatrixCoefficients_L1_negative"
+                    - "dsGetColorDepth_L1_positive"
+                    - "dsGetColorDepth_L1_negative"
+                    - "dsGetColorSpace_L1_positive"
+                    - "dsGetColorSpace_L1_negative"
+                    - "dsGetQuantizationRange_L1_positive"
+                    - "dsGetQuantizationRange_L1_negative"
+                    - "dsGetCurrentOutputSettings_L1_positive"
+                    - "dsGetCurrentOutputSettings_L1_negative"
+                    - "dsIsOutputHDR_L1_positive"
+                    - "dsIsOutputHDR_L1_negative"
+                    - "dsResetOutputToSDR_L1_positive"
+                    - "dsResetOutputToSDR_L1_negative"
+                    - "dsSetHdmiPreference_L1_positive"
+                    - "dsSetHdmiPreference_L1_negative"
+                    - "dsGetHdmiPreference_L1_positive"
+                    - "dsGetHdmiPreference_L1_negative"
+                    - "dsGetIgnoreEDIDStatus_L1_positive"
+                    - "dsGetIgnoreEDIDStatus_L1_negative"
+                    - "dsSetBackgroundColor_L1_positive"
+                    - "dsSetBackgroundColor_L1_negative"
+                    - "dsSetForceHDRMode_L1_positive"
+                    - "dsSetForceHDRMode_L1_negative"
+                    - "dsColorDepthCapabilities_L1_positive"
+                    - "dsColorDepthCapabilities_L1_negative"
+                    - "dsGetPreferredColorDepth_L1_positive"
+                    - "dsGetPreferredColorDepth_L1_negative"
+                    - "dsSetPreferredColorDepth_L1_positive"
+                    - "dsSetPreferredColorDepth_L1_negative"
             1:
                 name: "L2 dsVideoPort - Sink"
                 tests:
+                    - "L2_EnableDisabledVideoPorts"
+                    - "L2_VerifyDisplayAndPortStatus"
+                    - "L2_RetrieveAndVerifySurroundModeCapabilities"
+                    - "L2_VerifySupportedTvResolutions"
+                    - "L2_GetHDRCapabilities"
+                    - "L2_GetHDCPStatus"
+                    - "L2_VerifyHDCPProtocolStatus"
+                    - "L2_SetAndGetHdmiPreference"
+                    - "L2_GetColorSpace"
+                    - "L2_GetColorDepth"
+                    - "L2_GetQuantizationRange"
+                    - "L2_GetMatrixCoefficients"
+                    - "L2_SetAndGetResolution_source"
+                    - "L2_SetAndGetPreferredColorDepth_source"
+                    - "L2_CheckColorDepthCapabilities_source"
             2:
                 name: "L3 dsVideoPort"
                 tests:
