--- host/tests/dsVideoPort_L3_Tests/dsVideoPort_L3_testSetup.yml
@@ -1,20 +1,0 @@
-# this file is generated by the testing suites, when a new test is modified, this configuration file will be filled out and updated.
-#location: https://github.com/rdkcentral/rdk-halif-test-device_settings/blob/develop/host/tests/dsVideoPort_L3_tests/
-
-dsVideoPort:  # Prefix must always exist
-  description: "dsVideoPort Device Settings test setup"
-  assets:
-    device:
-      test1_Verify_Video_Content_Formats_Callbacks:
-        artifacts: #list of artifacts to be copied to device
-        execute: #list of commands to be executed on device
-        streams: #list of streams to be copied to device
-      test2_Check_DisplayConnected:
-        artifacts:
-        execute:
-        streams:
-
-
-
-
-
